Updates to send a limited size message to telegram

+++ src/globals.py
@@ -24,6 +24,7 @@ import urllib
 import numpy as np
 import pandas as pd
 from datetime import datetime
+import classes.PortfolioTracker as tracker
 from time import sleep
 from tabulate import tabulate
 from Telegram import send_message, send_photo, send_document, is_token_telegram_configured
@@ -469,11 +470,11 @@ def main(testing=False, testBuild=False, downloadOnly=False, startupoptions=None
                     return
             else:
                 if not downloadOnly:
-                    menuChoiceHierarchy = f'({selectedChoice["0"]}) {level0MenuDict[selectedChoice["0"]].strip()} > ({selectedChoice["1"]}) {level1MenuDict[selectedChoice["1"]].strip()} > ({selectedChoice["2"]}) {level2MenuDict[selectedChoice["2"]].strip()}'
+                    menuChoiceHierarchy = f'({selectedChoice["0"]}){level0MenuDict[selectedChoice["0"]].strip()}>({selectedChoice["1"]}){level1MenuDict[selectedChoice["1"]].strip()}>({selectedChoice["2"]}){level2MenuDict[selectedChoice["2"]].strip()}'
                     if selectedChoice['2'] == '6':
-                        menuChoiceHierarchy = menuChoiceHierarchy + f' > ({selectedChoice["3"]}) {level3ReversalMenuDict[selectedChoice["3"]].strip()}'
+                        menuChoiceHierarchy = menuChoiceHierarchy + f'>({selectedChoice["3"]}){level3ReversalMenuDict[selectedChoice["3"]].strip()}'
                     elif selectedChoice['2'] == '7':
-                        menuChoiceHierarchy = menuChoiceHierarchy + f' > ({selectedChoice["3"]}) {level3ChartPatternMenuDict[selectedChoice["3"]].strip()}'
+                        menuChoiceHierarchy = menuChoiceHierarchy + f'>({selectedChoice["3"]}){level3ChartPatternMenuDict[selectedChoice["3"]].strip()}'
                     print(colorText.BOLD + colorText.FAIL + '[+] You chose: ' + menuChoiceHierarchy + colorText.END)
                 listStockCodes = fetcher.fetchStockCodes(tickerOption, proxyServer=proxyServer, stockCode=None)
         except urllib.error.URLError as e:
@@ -614,25 +615,28 @@ def main(testing=False, testBuild=False, downloadOnly=False, startupoptions=None
             Utility.tools.clearScreen()
             screenResults = screenResults[screenResults['MA-Signal'].str.contains('Unknown') == False]
             screenResults = screenResults[screenResults[f'Trend({configManager.daysToLookback}Prds)'].str.contains('Unknown') == False]
-            menuChoiceHierarchy = menuChoiceHierarchy + ' (Data Period: ' + configManager.period + ', Candle Duration: ' + configManager.duration + ')'
+            menuChoiceHierarchy = menuChoiceHierarchy + f'({configManager.period} period, {configManager.duration} candles.)'
             print(colorText.BOLD + colorText.FAIL + '[+] You chose: ' + menuChoiceHierarchy + colorText.END)
             tabulated_results = tabulate(screenResults, headers='keys', tablefmt='grid')
             print(tabulated_results)
             if len(screenResults) >= 1:
                 if not testing:
-                    readyForPhoto = saveResults[saveResults['MA-Signal'].str.contains('Unknown') == False]
-                    readyForPhoto = readyForPhoto[readyForPhoto[f'Trend({configManager.daysToLookback}Prds)'].str.contains('Unknown') == False]
-                    markdown_results = tabulate(readyForPhoto, headers='keys', tablefmt='grid')
-                    pngName = 'PKScreener-result_' + \
-                            datetime.now().strftime("%d-%m-%y_%H.%M.%S")+".png"
-                    if is_token_telegram_configured():
-                        Utility.tools.tableToImage(markdown_results,tabulated_results,pngName,menuChoiceHierarchy)
-                        sendMessageToTelegramChannel(message=None, photo_filePath=pngName, caption=menuChoiceHierarchy)
-                        try:
-                            os.remove(pngName)
-                        except Exception as e:
-                            default_logger().debug(e, exc_info=True)
-                            pass
+                    if len(screenResults) <= 100:
+                        # No point sending a photo with more than 50 stocks.
+                        readyForPhoto = saveResults[saveResults['MA-Signal'].str.contains('Unknown') == False]
+                        readyForPhoto = readyForPhoto[readyForPhoto[f'Trend({configManager.daysToLookback}Prds)'].str.contains('Unknown') == False]
+                        menuChoiceHierarchy = f'({len(readyForPhoto)} stocks found). ' + menuChoiceHierarchy
+                        markdown_results = tabulate(readyForPhoto, headers='keys', tablefmt='grid')
+                        pngName = 'PKScreener-result_' + \
+                                datetime.now().strftime("%d-%m-%y_%H.%M.%S")+".png"
+                        if is_token_telegram_configured():
+                            Utility.tools.tableToImage(markdown_results,tabulated_results,pngName,menuChoiceHierarchy)
+                            sendMessageToTelegramChannel(message=None, photo_filePath=pngName, caption=menuChoiceHierarchy)
+                            try:
+                                os.remove(pngName)
+                            except Exception as e:
+                                default_logger().debug(e, exc_info=True)
+                                pass
                     print(colorText.BOLD + colorText.GREEN +
                             f"[+] Found {len(screenResults)} Stocks." + colorText.END)
                     Utility.tools.setLastScreenedResults(screenResults)
